+++ refiner/tests/integration/conftest.py
@@ -2,7 +2,21 @@ import os
 from pathlib import Path
 
 import pytest
+import pytest_asyncio
+from httpx import AsyncClient
 from testcontainers.compose import DockerCompose
+
+
+@pytest.fixture
+def auth_cookie():
+    return {"refiner-session": "test-token"}
+
+
+@pytest_asyncio.fixture
+async def authed_client(auth_cookie, base_url):
+    async with AsyncClient(base_url=base_url) as client:
+        client.cookies.update(auth_cookie)
+        yield client
 
 
 @pytest.fixture(scope="session", autouse=True)
